Move revoke permission tests to separate module.
